samples: basic: minimal: add no-sw-isr-table option
Add option to build without software isr table for platforms which
support it (ARM and RISCV). The software ISR table takes up
kilobytes of RAM and introduces latency, and it is not needed if
CONFIG_MULTITHREADING=n.

The initially supported nrf54l15 and nrf54lm20 boards additionally
need CONFIG_CLOCK_CONTROL=n for the minimal sample so board configs
have been added for them. Note that the entry in sample.yaml uses
EXTRA_CONF_FILE to not overwrite the board specific CONF_FILE.

Signed-off-by: Bjarki Arge Andreasen <bjarki.andreasen@nordicsemi.no>