soc: stm32: stm32h7rsx: add MPU region #0 disabling all accesses
Add a first region in the MPU to disable all access to the whole
memory range.  With that ensure that the MPU will block all
access to regions that aren't defined in further regions.
Ensure as well that the peripheral area is accessible.

This handles the errata 2.1.1 PLD might perform linefill to address
that would generate a MemManage Fault of the STM32H7Rxx / STM32H7Sxx
device errate ES0596 - Rev 6.

Signed-off-by: Alain Volmat <alain.volmat@foss.st.com>