doc: fs: document preconditions with fs_open & fs_opendir
fs_open and fs_opendir rely on the structures in the first param
being intialized. (The init functions are documented with the type
declarations.) This simple documentation change mentions this need
to initialize (and emphasizes it). (Users of similar APIs such as
POSIX open and opendir do not expect this initialization need.)

Fixes #97205
Signed-off-by: Paul Hepworth <paulh@vpitech.com>