ci: doc-build.yml: do not tweak manifest.project-filter unnecessarily
The documentation build shouldn't need to pull in anything beyond what's
active by default in the west.yml manifest. In fact, due to the way the
manifest_projects_table.py script currently works, it is important that
have a local west config that's "vanilla" so that the table of
active/inactive projects is correct.

Signed-off-by: Benjamin Cabé <benjamin@zephyrproject.org>