scripts: runners: nrf: Generalize the use of --erase-mode for all ICs
Until now, for historical reasons (see
a6ad26b51dd8f2dae519916e44408cb304856e45 and
ea20f759d4e5a71ecac2e721a293d4f95fa94ce1), the use of the --erase-mode
command-line switch was reserved for the nRF54L family. But in fact this
can be used (instead of --erase) for any of the Nordic ICs. This patch
extends the usage of this switch regardless of family.

Signed-off-by: Carles Cufi <carles.cufi@nordicsemi.no>